requirements.txt
requests
tzdata
supabase
python-dotenv